se agrego configuracion de eslint - prettier y husky

+++ src/app/core/guards/auth.guard.ts
@@ -1,43 +1,50 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, CanLoad, Route, Router, RouterStateSnapshot, UrlSegment, UrlTree } from '@angular/router';
+import {
+  ActivatedRouteSnapshot,
+  CanActivate,
+  CanLoad,
+  Route,
+  Router,
+  RouterStateSnapshot,
+  UrlSegment,
+  UrlTree,
+} from '@angular/router';
 import { Observable } from 'rxjs';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class AuthGuard implements CanActivate, CanLoad {
-
-  constructor(private router: Router){
-
-    console.log(router)
+  constructor(private router: Router) {
+    console.log(router);
   }
 
   canActivate(
     route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    state: RouterStateSnapshot
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    if (state.url.includes('?code=')) localStorage.clear();
 
-      if(state.url.includes("?code="))
-        localStorage.clear();
-        
-      if(localStorage.getItem('token'))
-      {
-        this.router.navigateByUrl('/');
-        return false;
-      }
-      else
-        return true;
-
+    if (localStorage.getItem('token')) {
+      this.router.navigateByUrl('/');
+      return false;
+    } else return true;
   }
   canLoad(
     route: Route,
-    segments: UrlSegment[]): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-      
-      if(localStorage.getItem('token'))
-      {
-        this.router.navigateByUrl('/');
-        return false;
-      }
-      else
-        return true;
+    segments: UrlSegment[]
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    if (localStorage.getItem('token')) {
+      this.router.navigateByUrl('/');
+      return false;
+    } else return true;
   }
 }
